N-Formyl Met is NOT a N-formylated residue
So... this is a weird one... since fMet is loaded by a tRNA onto the ribosome, MOD:00030, which the natural, nonstandard encoded residue (MOD:00868) is NOT N-formylated, because it's not a PTM to a methionine that's loaded by the ribosome, but rather the entity (like any other AA) that is loaded onto the ribosome.

Note that MOD:00482 is the entity that refers to a formyl modification to a methionine!